WIP: Add demo showing how to write internal tests
Add and test a utility, `is_internal`, that tells whether Pulp Smash is
on the internal network or now. Add a skeleton module,
`pulp-smash.tests.rpm.api_v2.test_internal`. This module demonstrates
how to write tests that only run on the internal network. (Naturally, it
uses the new function.)

To run the tests and prove to yourself that it works, configure an
environment as described in the README and elsewhere, and execute:

    python -m unittest2 pulp_smash.tests.rpm.api_v2.test_internal

Or:

    py.test pulp_smash/tests/rpm/api_v2/test_internal.py

Or whatever other test runner you fancy.